Removed logic from TwigExtension
* Added new service
* BC Break in Twig service (not supporting BC Compat there)
* Added some functional testing in managers
* Updated all dependencies with Twig to 1.23 (for deprecation resolving)
* TwigExtension now implementing interface for Globals (for deprecation resolving)
