+++ src/Elcodi/Bundle/LanguageBundle/Resources/config/twig.yml
@@ -6,7 +6,6 @@ services:
     elcodi.twig_extension.language:
         class: Elcodi\Component\Language\Twig\LanguageExtension
         arguments:
-            - @elcodi.manager.language
-            - @elcodi.locale
+            - @elcodi.manager.promoted_language
         tags:
             - { name: twig.extension }
