+++ src/Elcodi/Component/Currency/composer.json
@@ -36,7 +36,7 @@
         "doctrine/orm": "^2.5",
         "symfony/http-foundation": "^2.7",
         "sebastian/money": "^1.5",
-        "twig/twig": "^1.18",
+        "twig/twig": "^1.23",
 
         "elcodi/core": "^1.0",
         "elcodi/language": "^1.0"
