+++ src/Elcodi/Component/Language/composer.json
@@ -34,7 +34,7 @@
         "php": "^5.4|^7.0",
         "doctrine/common": "^2.5",
         "doctrine/orm": "^2.5",
-        "twig/twig": "^1.18",
+        "twig/twig": "^1.23",
 
         "elcodi/core": "^1.0"
     },
