+++ src/Elcodi/Component/Metric/composer.json
@@ -34,7 +34,7 @@
         "doctrine/common": "^2.5",
         "doctrine/orm": "^2.5",
         "symfony/console": "^2.7",
-        "twig/twig": "^1.18",
+        "twig/twig": "^1.23",
         "predis/predis": "^0.8.7",
 
         "elcodi/core": "^1.0"
