+++ src/Elcodi/Component/Payment/composer.json
@@ -33,7 +33,7 @@
     "require": {
         "php": "^5.4|^7.0",
         "symfony/event-dispatcher": "^2.7",
-        "twig/twig": "^1.18",
+        "twig/twig": "^1.23",
 
         "elcodi/core": "^1.0"
     },
