+++ src/Elcodi/Component/Store/Exception/StoreNotFoundException.php
@@ -27,11 +27,11 @@ class StoreNotFoundException extends Exception
     /**
      * (PHP 5 &gt;= 5.1.0)<br/>
      * Construct the exception. Note: The message is NOT binary safe.
-     * 
+     *
      * @param string    $message  [optional] The Exception message to throw.
      * @param int       $code     [optional] The Exception code.
      * @param Exception $previous [optional] The previous exception used for the exception chaining. Since 5.3.0
-     * 
+     *
      * @link http://php.net/manual/en/exception.construct.php
      */
     public function __construct($message = "", $code = 0, Exception $previous = null)
